searchtree

+++ bigNumber.h
@@ -3,8 +3,8 @@
 #include <string>
 using namespace std;
 
-typedef unsigned short BASE;
-typedef unsigned int DBASE;
+typedef unsigned int BASE;
+typedef unsigned long long DBASE;
 
 #define BASE_SIZE  (sizeof(BASE) * 8)
 #define BASE_VAL   ((DBASE)1 << BASE_SIZE)
@@ -42,7 +42,6 @@ public:
     BigNumber  operator*(const BigNumber& bn) const;
     BigNumber& operator*=(const BigNumber& bn);
     BigNumber  operator*(BASE v) const;
-    // BigNumber operator*(DBASE v) const;
     BigNumber& operator*=(BASE v);
 
     void divBn(const BigNumber& bn, BigNumber& quotient, BigNumber& remainder) const;
